feat(corfu): vim-like C-n / C-p completion
Adds insert-mode bindings to make C-n and C-p trigger the Corfu popup,
with the first or last candidate selected, respectively.

Care has been taken that no candidate is actually inserted, even if
there is only one candidate - only the popup will be shown with the
candidate selected. Whether the candidate is previewed or not depends on
`corfu-preview-current`.